Move get_common_config logic into TaskConfigHelper
