+++ core/task_runner.py
@@ -61,13 +61,13 @@ class TaskRunner:
     def initialize_tasks(self) -> List[BaseTask]:
         """Initialize all enabled tasks from configuration"""
         tasks = []
-        common_config = BaseTask.get_common_config()
+        common_config = TaskConfigHelper.get_common_config()
+        global_task_class_path = self.global_config.get("task_class")
+        global_config_values = self.global_config.get("config", {})
 
         self.initialize_hummingbot_client_config()
         self.enable_vpn_compatibility()
 
-        global_task_class_path = self.global_config.get("task_class")
-        global_config_values = self.global_config.get("config", {})
         for task_name, task_config in self.tasks_config.items():
             if not task_config.get("enabled", True):
                 logger.info(f"Skipping disabled task: {task_name}")
